Move UDPPeerGroup into a separate file.
