+++ src/io/iohk/scalanet/peergroup/PeerGroup.scala
@@ -1,6 +1,5 @@
 package io.iohk.scalanet.peergroup
 
-import java.net.InetSocketAddress
 import java.nio.ByteBuffer
 
 import cats.data.Kleisli
@@ -18,28 +17,5 @@ object PeerGroup {
   type Lift[F[_]] = Kleisli[F, Task[Unit], Unit]
 
   abstract class TerminalPeerGroup[A, F[_]] extends PeerGroup[A, F]
-
-  class UDPPeerGroup[F[_]](implicit liftF: Lift[F]) extends TerminalPeerGroup[InetSocketAddress, F]() {
-
-    // TODO start listening
-    println("UDPPeerGroup starting")
-
-    override def sendMessage(address: InetSocketAddress, message: ByteBuffer): F[Unit] = {
-      val send: Task[Unit] = Task {
-        println(s"UDPPeerGroup, send to address $address, message $message")
-        // TODO send the message
-        ()
-      }
-      liftF(send)
-    }
-
-    override def shutdown(): F[Unit] = {
-      liftF(Task {
-        println(s"UDPPeerGroup, shutdown")
-        // TODO shutdown
-        ()
-      })
-    }
-  }
 }
