+++ src/io/iohk/scalanet/peergroup/eithert.scala
@@ -17,7 +17,7 @@ object eithert {
 
   type TerminalPeerGroup[A] = io.iohk.scalanet.peergroup.PeerGroup.TerminalPeerGroup[A, ET]
 
-  type UDPPeerGroup = io.iohk.scalanet.peergroup.PeerGroup.UDPPeerGroup[ET]
+  type UDPPeerGroup = io.iohk.scalanet.peergroup.UDPPeerGroup[ET]
 
   implicit def liftEitherT(implicit ec: ExecutionContext): Lift[ET] =
     Kleisli { task: Task[Unit] =>
